+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## Unreleased
-- Replaced the GTK-based popup menu and dialogs with a pure-SDL implementation. On Android, the popup menu opens with a three-finger tap.
+- Replaced the GTK-based popup menu and dialogs with a pure-SDL implementation.
+- On Android, the popup menu opens with a three-finger tap.
+- Implemented volume control panel.
+- Added `mute_on_unfocus` option to mute audio while the window is unfocused.
 
 ## 2.18.0 - 2026-06-23
 - Added support for the intro demo of Daiakuji.